requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aws-assume-role-lib==2.10.0
    # via -r requirements.in
bagit==1.8.1
    # via -r requirements.in
boto3==1.28.57
    # via
    #   -r requirements.in
    #   aws-assume-role-lib
botocore==1.31.57
    # via
    #   boto3
    #   s3transfer
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
pillow==10.2.0
    # via -r requirements.in
pypdf==4.1.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
s3transfer==0.7.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.16
    # via botocore
